Renamed databases to clients
