+++ lib/clients/elasticsearch.js
