+++ lib/clients/redis.js
