+++ lib/models/model_Es_test.js
@@ -7,7 +7,7 @@
 // Dependencies
 var Config = require('../config/config');
 var _ = require('lodash');
-var es = require('../databases/elasticsearch')(Config);
+var es = require('../clients/elasticsearch')(Config);
 
 // Export
 module.exports = new Elasticsearch();
