+++ lib/models/model_Mysql_test.js
@@ -7,7 +7,7 @@
 // Dependencies
 var Config = require('../config/config');
 var _ = require('lodash');
-var mysql = require('../databases/mysql')(Config);
+var mysql = require('../clients/mysql')(Config);
 
 // Export
 module.exports = new Mysql_test();
