+++ lib/models/model_Redis_test.js
@@ -7,7 +7,7 @@
 // Dependencies
 var Config = require('../config/config');
 var _ = require('lodash');
-var redis = require('../databases/redis')(Config);
+var redis = require('../clients/redis')(Config);
 
 // Export
 module.exports = new Redis();
